requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.12.13
    # via alpaca-trade-api
aiosignal==1.3.2
    # via aiohttp
alpaca-trade-api==3.2.0
    # via alpaca-tui-trader (pyproject.toml)
attrs==25.3.0
    # via aiohttp
certifi==2025.6.15
    # via requests
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via alpaca-tui-trader (pyproject.toml)
colorama==0.4.6
    # via alpaca-tui-trader (pyproject.toml)
deprecation==2.1.0
    # via alpaca-trade-api
frozenlist==1.7.0
    # via
    #   aiohttp
    #   aiosignal
idna==3.10
    # via
    #   requests
    #   yarl
linkify-it-py==2.0.3
    # via markdown-it-py
markdown-it-py==3.0.0
    # via
    #   mdit-py-plugins
    #   rich
    #   textual
mdit-py-plugins==0.4.2
    # via markdown-it-py
mdurl==0.1.2
    # via markdown-it-py
msgpack==1.0.3
    # via alpaca-trade-api
multidict==6.6.0
    # via
    #   aiohttp
    #   yarl
numpy==2.3.1
    # via
    #   alpaca-trade-api
    #   pandas
packaging==25.0
    # via deprecation
pandas==2.3.0
    # via alpaca-trade-api
platformdirs==4.3.8
    # via textual
propcache==0.3.2
    # via
    #   aiohttp
    #   yarl
pygments==2.19.2
    # via rich
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.1.1
    # via alpaca-tui-trader (pyproject.toml)
pytz==2025.2
    # via pandas
pyyaml==6.0.1
    # via alpaca-trade-api
requests==2.32.4
    # via alpaca-trade-api
rich==14.0.0
    # via textual
six==1.17.0
    # via python-dateutil
textual==3.5.0
    # via alpaca-tui-trader (pyproject.toml)
typing-extensions==4.14.0
    # via textual
tzdata==2025.2
    # via pandas
uc-micro-py==1.0.3
    # via linkify-it-py
urllib3==1.26.20
    # via
    #   alpaca-trade-api
    #   requests
websocket-client==1.8.0
    # via alpaca-trade-api
websockets==10.4
    # via alpaca-trade-api
yarl==1.20.1
    # via aiohttp
